requirements.txt
torch
torchvision
numpy
matplotlib
opencv_python
Pillow
tqdm
mxnet
bcolz
easydict
scikit_learn
scipy

